Add method to get organization of currently logged in user
